finish

+++ Common/helper.h
@@ -3,11 +3,11 @@
 
 static unsigned long long seed = 1;
 
-double drand48()
+float drand48()
 {
     seed = (0x5DEECE66DLL * seed + 0xB16) & 0xFFFFFFFFFFFFLL;
     unsigned long long x = seed >> 16;
-    return  ((double)x / (double)0x100000000LL);
+    return  (float)((double)x / (double)0x100000000LL);
 }
 
 void srand48(unsigned int i)
